Deploy Team Member Management

See merge request obviobysage/event/web!40

+++ src/organization/OrganizationProvider.tsx
@@ -4,11 +4,12 @@ import {api, createRoutes} from 'lib/url'
 import {Organization} from 'organization'
 import {organizationTokenKey} from 'organization/auth'
 import {obvioClient} from 'obvio/obvio-client'
-import React, {useCallback} from 'react'
+import React, {useCallback, useMemo} from 'react'
 import {useLocation} from 'react-router-dom'
 
-export type OrganizationClient = ReturnType<typeof createClientFor>
+export type OrganizationClient = NonNullable<ReturnType<typeof useClient>>
 export type OrganizationRoutes = ReturnType<typeof createRoutesFor>
+
 interface OrganizationContextProps {
   organization: Organization
   routes: OrganizationRoutes
@@ -29,12 +30,13 @@ export default function OrganizationProvider(props: {
     return findOrganization(slug)
   }, [slug])
   const {data: organization, loading} = useAsync(find)
+  const client = useClient(organization)
 
   if (loading) {
     return null
   }
 
-  if (!organization) {
+  if (!organization || !client) {
     return (
       <div>
         <h1>404 - Organization '{slug}' not found.</h1>
@@ -43,7 +45,6 @@ export default function OrganizationProvider(props: {
   }
 
   const routes = createRoutesFor(organization)
-  const client = createClientFor(organization)
 
   return (
     <OrganizationContext.Provider
@@ -73,6 +74,7 @@ export function createRoutesFor(organization: Organization) {
   return createRoutes(
     {
       login: '/login',
+      team: '/team',
       events: {
         create: '/create',
         ':event': {
@@ -86,18 +88,24 @@ export function createRoutesFor(organization: Organization) {
   )
 }
 
-function createClientFor(organization: Organization): typeof client {
-  const tokenKey = organizationTokenKey(organization.slug)
-  return {
-    get: (url: string, options?: RequestOptions) =>
-      client.get(url, {...options, tokenKey}),
-    post: (url: string, data: {}, options?: RequestOptions) =>
-      client.post(url, data, {...options, tokenKey}),
-    put: (url: string, data: {}, options?: RequestOptions) =>
-      client.put(url, data, {...options, tokenKey}),
-    delete: (url: string, options?: RequestOptions) =>
-      client.delete(url, {...options, tokenKey}),
-  }
+function useClient(organization: Organization | null): typeof client | null {
+  return useMemo(() => {
+    if (!organization) {
+      return null
+    }
+
+    const tokenKey = organizationTokenKey(organization.slug)
+    return {
+      get: (url: string, options?: RequestOptions) =>
+        client.get(url, {...options, tokenKey}),
+      post: (url: string, data: {}, options?: RequestOptions) =>
+        client.post(url, data, {...options, tokenKey}),
+      put: (url: string, data: {}, options?: RequestOptions) =>
+        client.put(url, data, {...options, tokenKey}),
+      delete: (url: string, options?: RequestOptions) =>
+        client.delete(url, {...options, tokenKey}),
+    }
+  }, [organization])
 }
 
 function findOrganization(slug: string) {
